requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
autopep8==1.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.21.42
botocore==1.24.42
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==36.0.2
daphne==3.0.2
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-rest-multiple-models==2.1.3
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.0
Pillow
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
whitenoise==5.3.0
zope.interface==5.4.0
